python-services/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.5
google-generativeai>=0.3.1
lancedb>=0.3.4
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.2.2
pydantic>=2.5.0
torch>=2.0.0